requirements.txt
python-dotenv
mysql-connector-python
pandas
pinecone
langchain-google-genai
ipywidgets
langchain-pinecone
streamlit
